FIX :: board랑 like class 분리
