+++ src/main/java/com/bamdoliro/gati/domain/board/domain/repository/BoardRepository.java
@@ -1,7 +1,7 @@
 package com.bamdoliro.gati.domain.board.domain.repository;
 
 import com.bamdoliro.gati.domain.board.domain.Board;
-import com.bamdoliro.gati.domain.board.domain.type.Status;
+import com.bamdoliro.gati.domain.board.domain.type.board.Status;
 import com.bamdoliro.gati.domain.community.domain.Community;
 import org.springframework.data.jpa.repository.JpaRepository;
