+++ src/main/java/com/bamdoliro/gati/domain/board/facade/BoardFacade.java
@@ -1,13 +1,10 @@
 package com.bamdoliro.gati.domain.board.facade;
 
 import com.bamdoliro.gati.domain.board.domain.Board;
-import com.bamdoliro.gati.domain.board.domain.repository.BoardLikeRepository;
 import com.bamdoliro.gati.domain.board.domain.repository.BoardRepository;
-import com.bamdoliro.gati.domain.board.domain.type.Status;
+import com.bamdoliro.gati.domain.board.domain.type.board.Status;
 import com.bamdoliro.gati.domain.board.exception.BoardNotFoundException;
-import com.bamdoliro.gati.domain.board.exception.LikeOverlapException;
 import com.bamdoliro.gati.domain.community.domain.Community;
-import com.bamdoliro.gati.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -18,8 +15,8 @@ import java.util.List;
 public class BoardFacade {
 
     private final BoardRepository boardRepository;
-    private final BoardLikeRepository boardLikeRepository;
 
+    // 삭제되지 않는 게시물 id 로 찾기
     public Board findBoardById(Long id) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> BoardNotFoundException.EXCEPTION);
@@ -28,14 +25,10 @@ public class BoardFacade {
         return board;
     }
 
+    // 공동체랑 삭제 여부로 찾기
     public List<Board> findByCommunityAndStatus(Community community, Status status) {
         return boardRepository.findByCommunityAndStatus(community, status)
                 .orElseThrow(() -> BoardNotFoundException.EXCEPTION);
     }
 
-    public void validateLike(Board board, User user) {
-        if (boardLikeRepository.existsByBoardAndLiker(board, user)) {
-            throw LikeOverlapException.EXCEPTION;
-        }
-    }
 }
